common: fix rt_ops setup in _scratch_mkfs_sized
Tests using _scratch_mkfs_sized() will fail if SCRATCH_RTDEV is set,
but, USE_EXTERNAL is not, this happens because the function pass
"-r size=$fssize" to the _scratch_mkfs_xfs argument line, which in turn
will not set rtdev because USE_EXTERNAL is not set.

Tests like xfs/015 will fail as:

xfs/015 6s ... [failed, exit status 1]- output mismatch
.
.
    +size specified for non-existent rt subvolume
    +Usage: mkfs.xfs
.
.

with this patch the test runs properly using the rtdev if USE_EXTERNAL
is set.

Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>