fstests: add stress truncation + writeback + compaction split test
Stress test folio splits by using the debugfs interface to a target
a new smaller folio order while running compaction at the same time.
This is dangerous at the moment as its using a debugfs API which
requires two out of tree fixes [0] [1] which have already been
posted but not yet merged.

With these debugfs patches applied this test can now be used to
reproduce an issue which was only possible to reproduce by running
generic/447 twice with min order:

https://gist.github.com/mcgrof/d12f586ec6ebe32b2472b5d634c397df

This is designed to try to exacerbate races with folio splits incurred
by truncation and race that with compaction and writeback. This only
creates a crash with min order enabled, so for example with a 16k block
sized XFS test profile.

This also begs the question if something like MADV_NOHUGEPAGE might be
desirable from userspace, so to enable userspace to request splits when
possible.

If inspecting more closely, you'll want to enable on your kernel boot:

	dyndbg='file mm/huge_memory.c +p'

Since we want to race large folio splits we also augment the full test
output log $seqres.full with the test specific number of successful
splits from vmstat thp_split_page and thp_split_page_failed.

[0] https://git.kernel.org/pub/scm/linux/kernel/git/mcgrof/linux.git/commit/?h=20240424-lbs&id=80f6df5037fd0ad560526af45bd7f4d779fe03f6
[1] https://git.kernel.org/pub/scm/linux/kernel/git/mcgrof/linux.git/commit/?h=20240424-lbs&id=38f6fac5b4283ea48b1876fc56728f062168f8c3
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>